+++ projects/ngx-ukis-ui-clarity/src/lib/global-alert/global-alert.component.spec.ts
@@ -1,6 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { GlobalAlertComponent } from './global-alert.component';
+import { AlertService, IAlert } from './alert.service';
+import { ClarityModule } from '@clr/angular';
 
 describe('GlobalAlertComponent', () => {
   let component: GlobalAlertComponent;
@@ -8,12 +10,18 @@ describe('GlobalAlertComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [GlobalAlertComponent]
+      imports: [ClarityModule, GlobalAlertComponent],
+      providers: [AlertService]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(GlobalAlertComponent);
     component = fixture.componentInstance;
+    component.alert = {
+      type: 'info',
+      text: 'test Alert',
+      closeable: true
+    } as IAlert;
     fixture.detectChanges();
   });