+++ projects/ngx-ukis-ui-clarity/src/lib/global-alert/global-alert.component.ts
@@ -1,11 +1,25 @@
-import { Component } from '@angular/core';
+import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { IAlert } from './alert.service';
+
+import { ClarityIcons, infoCircleIcon, windowCloseIcon } from '@cds/core/icon';
+import { NgClass } from '@angular/common';
+import { ClrIconModule, ClrAlertModule } from '@clr/angular';
+ClarityIcons.addIcons(...[infoCircleIcon, windowCloseIcon]);
 
 @Component({
-  selector: 'ngx-global-alert',
-  imports: [],
+  selector: 'ngx-ukis-global-alert',
   templateUrl: './global-alert.component.html',
-  styleUrl: './global-alert.component.scss'
+  styleUrl: './global-alert.component.scss',
+  imports: [NgClass, ClrIconModule, ClrAlertModule]
 })
 export class GlobalAlertComponent {
+  @Input() alert!: null | IAlert;
+  @Output() alertChange = new EventEmitter();
+  constructor() { }
+
+  close() {
+    this.alert = null;
+    this.alertChange.emit(this.alert);
+  }
 
 }