+++ projects/ngx-ukis-ui-clarity/src/lib/global-progress/global-progress.component.spec.ts
@@ -1,6 +1,8 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { GlobalProgressComponent } from './global-progress.component';
+import { ProgressService, IProgress } from './progress.service';
+import { ClarityModule } from '@clr/angular';
 
 describe('GlobalProgressComponent', () => {
   let component: GlobalProgressComponent;
@@ -8,12 +10,16 @@ describe('GlobalProgressComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [GlobalProgressComponent]
+      imports: [ClarityModule, GlobalProgressComponent],
+      providers: [ProgressService]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(GlobalProgressComponent);
     component = fixture.componentInstance;
+    component.progress = {
+      indeterminate: true
+    } as IProgress;
     fixture.detectChanges();
   });